Begin separating global set_instructions from operation-specific set-instructions with the filtering operation
