+++ lib/active_set/enumerable_set_instruction.rb
@@ -9,40 +9,27 @@ class ActiveSet
     end
 
     def attribute_value_for(item)
-      item_value = @attribute_instruction
-                   .value_for(item: item)
-      item_value = item_value.downcase if case_insensitive_operation_for?(item_value)
-      item_value
+      @item_values ||= Hash.new do |h, key|
+        item_value = @attribute_instruction.value_for(item: key)
+        item_value = item_value.downcase if case_insensitive_operation_for?(item_value)
+        h[key] = item_value
+      end
+
+      @item_values[item]
     end
 
-    # rubocop:disable Lint/UnderscorePrefixedVariableName
-    def attribute_value
-      _attribute_value = @attribute_instruction.value
-      _attribute_value = _attribute_value.downcase if case_insensitive_operation_for?(_attribute_value)
-      _attribute_value
+    def instruction_value
+      return @instruction_value if defined? @instruction_value
+
+      instruction_value = @attribute_instruction.value
+      instruction_value = instruction_value.downcase if case_insensitive_operation_for?(instruction_value)
+      @instruction_value = instruction_value
     end
-    # rubocop:enable Lint/UnderscorePrefixedVariableName
 
     def case_insensitive_operation_for?(value)
       return false unless @attribute_instruction.case_insensitive?
 
       value.is_a?(String) || value.is_a?(Symbol)
-    end
-
-    def attribute_instance
-      set_item = @set.find(&:present?)
-      return set_item if @attribute_instruction.associations_array.empty?
-      return @attribute_model if defined? @attribute_model
-
-      @attribute_model = @attribute_instruction
-                         .associations_array
-                         .reduce(set_item) do |obj, assoc|
-        obj.public_send(assoc)
-      end
-    end
-
-    def attribute_class
-      attribute_instance&.class
     end
   end
 end
