+++ lib/active_set/filtering/active_record/set_instruction.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require_relative '../../active_record_set_instruction'
+
+class ActiveSet
+  module Filtering
+    module ActiveRecord
+      class SetInstruction < ActiveRecordSetInstruction
+      end
+    end
+  end
+end
